Added post policy

+++ app/Http/Controllers/BlogController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Http\Requests\CreatePostRequest;
+use App\Models\Post;
 use App\Post\PostService;
 
 class BlogController extends Controller
@@ -17,11 +18,13 @@ class BlogController extends Controller
 
     public function getCreatePostForm()
     {
+        $this->authorize('create', Post::class);
         return view('panel.create_post');
     }
 
     public function createPost(CreatePostRequest $request)
     {
+        $this->authorize('create', Post::class);
         $this->postService->createPost($request->validated());
         return redirect()->route('home');
     }
@@ -34,17 +37,20 @@ class BlogController extends Controller
 
     public function getEditPostForm(int $postId)
     {
+        $this->authorize('edit', Post::class);
         return view('panel.create_post', ["post" => $this->postService->getPost($postId)]);
     }
 
     public function editPost(CreatePostRequest $request, int $postId)
     {
+        $this->authorize('edit', Post::class);
         $this->postService->editPost($postId, $request->validated());
         return redirect()->route('blog.postlist');
     }
 
     public function deletePost(int $postId)
     {
+        $this->authorize('delete', Post::class);
         $this->postService->deletePost($postId);
         return redirect()->route('blog.postlist');
     }
